fix : - put move and new_position in new file movements.c and added move_angle for norm - added vars rot_s and mov_s to change the speed according to the screen width
